refactor: BattleHeader 설정 메뉴 분리로 리렌더 최소화 (#258) - BattleSettingsMenu 컴포넌트로 설정/사운드 로직 분리 - BattleHeader에서 설정 상태 제거 - 토글/볼륨 변경 시 헤더 전체 리렌더 방지
